Use Serde instead of "manual" work
- Implement `Visitor`, `Deserialize` and `Serialize` for `Pieces`
- Implement `Visitor`, `Deserialize` and `Serialize` for `Peers`
- Remove `hash.rs`
- Rename `peers.rs` to `tracker.rs`

+++ src/constants.rs
@@ -1,14 +1,17 @@
+/// Length of the used hash sum, which is the [`sha1::Sha1`] sum, and its length is always 20
 pub const SHA1_LEN: usize = 20;
 
 /// Our own Peer ID
 ///
 /// https://wiki.theory.org/BitTorrentSpecification#peer_id
+///
+/// We fixed an arbitrary value here.
 pub const PEER_ID: &str = "AAA3AAAAAAA7AAAABAAA";
 
-pub const PORT: usize = 6881;
+pub const PORT: u16 = 6881;
 pub const UPLOADED: usize = 0;
 pub const DOWNLOADED: usize = 0;
 
 /// https://www.bittorrent.org/beps/bep_0023.html
-pub const COMPACT: usize = 1;
+pub const COMPACT: u8 = 1;
 pub const PEER_LEN: usize = 6;
